requirements.txt
Flask
chromadb
requests
python-dotenv
PyPDF2
python-docx
chardet
SQLAlchemy 